feat(storage-azure): expose storage client
